Create Main.java
This is paid educational game in form of an App where people all over the world commit some sum of money in order to play the game. If the player is versed in the concerned topic he wins more than he has committed otherwise,he gets kicked in the butt and tempted to play more after carrying out his research.The game encourages proper study. It is planned to be  world wide but we are starting with Nigerian history.